pull to

+++ modules/activiti-engine/src/main/java/org/activiti/engine/impl/form/LongFormType.java
@@ -23,7 +23,7 @@ import org.activiti.engine.form.AbstractFormType;
 public class LongFormType extends AbstractFormType {
 
   public String getName() {
-    return "double";
+    return "long";
   }
 
   public String getMimeType() {
@@ -34,7 +34,7 @@ public class LongFormType extends AbstractFormType {
     if (propertyValue==null || "".equals(propertyValue)) {
       return null;
     }
-    return new Double(propertyValue);
+    return new Long(propertyValue);
   }
 
   public String convertModelValueToFormValue(Object modelValue) {
